+++ ak_library_management/models/borrow_transaction_history.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from datetime import date
+from datetime import date, timedelta
 from odoo import models, fields, api
 from odoo.exceptions import ValidationError
 
@@ -62,22 +62,48 @@ class BorrowTransactionHistory(models.Model):
         book return date.
         """
         borrow_transaction_book_ids = self.search([('borrow_end_date', '<=', date.today()),
-                                 ('books.status', '=', 'borrowed')])
+                                                   ('books.status', '=', 'borrowed')])
         for rec in borrow_transaction_book_ids:
             mail_template = self.env.ref(
                 'ak_library_management.email_template_library_book_reminder')
             mail_template.send_mail(rec.id, force_send=True)
 
-    def automated_action_overdue_books(self):
+    def action_overdue_books(self):
         """
         This method used for customer cant borrow book without returning old books
         which is overdue of return date.
         """
-        recs = self.search([('customer_id.id', "=", self.customer_id.id)])
-        for rec in recs:
+        borrow_transaction_history_id = self.search([('customer_id.id', "=", self.customer_id.id)])
+        for rec in borrow_transaction_history_id:
             for book in rec.books:
                 if rec.borrow_end_date < date.today() and book.status == "borrowed":
                     raise ValidationError(f"{rec.customer_id.name}"
                                           f"with overdue books"
                                           f"cannot borrow new ones until"
                                           f"you return the overdue items.")
+
+    def action_book_returned_reminder_days(self):
+        """
+        this scheduled action method used for reminding users to book returned date.
+        """
+        reminder_date = date.today() + timedelta(days=2)
+        records = self.search([
+            ('borrow_end_date', '=', reminder_date),
+            ('books.status', '=', 'borrowed')
+        ])
+        for record in records:
+            self.env['bus.bus']._sendone(record.customer_id, 'simple_notification', {
+                'type': 'warning',
+                'message': f"{record.customer_id.name} your book return date is after 2 days, please return it before due date",
+            })
+
+    def action_mark_books_as_returned(self):
+        """
+        this server action method used for mark books as returned.
+        """
+        for book in self.books:
+            book.status = "returned"
+            self.env['bus.bus']._sendone(self.customer_id, 'simple_notification', {
+                'type': 'warning',
+                'message': f"{self.customer_id.name} your return {book.name} book has been recorded",
+            })